// src/types/Trimestre.ts
export class Trimestre {
  date: Date;
  nbrMoisGrossesse: string;
  tensionArterielle: string;
  ta1: boolean;
  ta2: boolean;
  ta3: boolean;
  ta1Meaning: string;
  ta2Meaning: string;
  ta3Meaning: string;
  anemieSigneClinique: boolean;
  proteineUrinaire: string;
  tauxHB: boolean;
  poids: string;
  hauteurUterine: string;
  hauteurUterineAnormale: boolean;
  mafBcf: boolean;
  palpation: string;
  presentationVicieuse: boolean;
  toucherVaginal: string;
  colOuvert: boolean;
  autreQueCephalique: boolean;
  perteSang: boolean;
  promontoireAccessible: boolean;
  observationSoins: string;
  echoId: string;
  nextRDV: string;
  prisePoidsAnormale: boolean;
  id: string;
  idPatient: string;
  idDossier: string;
  position: number;
  constructor(idPatient: string, idDossier: string, position: number) {
    this.date = new Date();
    this.nbrMoisGrossesse = "";
    this.tensionArterielle = "";
    this.ta1 = false;
    this.ta2 = false;
    this.ta3 = false;
    this.ta1Meaning = "T.A. > 14/09";
    this.ta2Meaning = "T.A. > 13/08 et Prot. ++ ou >";
    this.ta3Meaning = "T.A. > 14/09 et Prot. ++ ou >";
    this.anemieSigneClinique = false;
    this.proteineUrinaire = "";
    this.tauxHB = false;
    this.poids = "";
    this.hauteurUterine = "";
    this.hauteurUterineAnormale = false;
    this.mafBcf = false;
    this.palpation = "";
    this.presentationVicieuse = false;
    this.toucherVaginal = "";
    this.colOuvert = false;
    this.autreQueCephalique = false;
    this.perteSang = false;
    this.promontoireAccessible = false;
    this.observationSoins = "";
    this.echoId = "";
    this.nextRDV = "";
    this.prisePoidsAnormale = false;
    this.id = "";
    this.idPatient = idPatient;
    this.idDossier = idDossier;
    this.position = position;
  }
}